@@ -4,21 +4,22 @@
   "private": true,
   "homepage": "https://ahusan.github.io",
   "dependencies": {
-    "@headlessui/react": "^1.7.4",
-    "@heroicons/react": "^2.0.13",
-    "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^12.1.3",
-    "@testing-library/user-event": "^13.5.0",
+    "@headlessui/react": "^1.7.14",
+    "@heroicons/react": "^2.0.17",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.4.3",
     "localforage": "^1.10.0",
     "match-sorter": "^6.3.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router": "^6.10.0",
-    "react-router-dom": "^6.10.0",
-    "react-scripts": "5.0.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-icons": "^4.8.0",
+    "react-router": "^6.11.0",
+    "react-router-dom": "^6.11.0",
+    "react-scripts": "5.0.1",
     "sort-by": "^1.2.0",
-    "tailwindcss": "^3.2.4",
-    "web-vitals": "^2.1.4"
+    "tailwindcss": "^3.3.2",
+    "web-vitals": "^3.3.1"
   },
   "scripts": {
     "predeploy": "npm run build",
